feat: add LabelUiState to main navigation drawer

+++ feature/mainscreen/src/main/java/com/mshdabiola/mainscreen/state/NoteType.kt
@@ -5,7 +5,7 @@ sealed class NoteType {
     object TRASH : NoteType()
     object ARCHIVE : NoteType()
     object REMAINDER : NoteType()
-    data class LABEL(val index: Int) : NoteType()
+    data class LABEL(val index: Long) : NoteType()
 }
 
 fun NoteType.toNoteType() = when (this) {
